playerHPバーの修復完了

+++ HuntingBrave_client/source/BaseChara.cpp
@@ -72,6 +72,8 @@
 	//	•`‰æ
 	void	BaseChara::Render( iexShader* shader, LPSTR technique )
 	{
+		
+
 		if ( shader == nullptr || technique == nullptr )
 		{
 			obj->Render();
@@ -80,7 +82,8 @@
 		{
 			obj->Render( shader, technique );
 		}
-		bar->Render( GetLifeInfo().life, GetPos(), GetUp() );
+
+		bar->Render((float)GetLifeInfo().life, GetPos(), GetUp());
 		//drawShape->DrawCapsule( collisionInfo.collisionShape.capsule.p1, collisionInfo.collisionShape.capsule.p2, collisionInfo.radius, 0xFFFFFFFF );
 	}
